once: add init_from_ref() method
Add a method to initialize a Once from a reference to a &T (which
implements Copy). This method can be used to avoid expensive copies into
the stack in debug builds, which might not be able to optimize away the
moving of the value into the internal cell.

Add as well a basic test to ensure the functionality is correct.